+++ Orbit/src/ui/UIStandard.h
@@ -10,7 +10,7 @@ public:
 
 	virtual void onUpdate() override;
 	virtual void show(bool show) override;
-	virtual float getValue() const override;
+	virtual int getValue() const override;
 	virtual UIType getType() const override;
 
 private:
